backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
